Cleanup procedural streaming code to use streaming controller with distance streamin-stream out

+++ Assets/ProceduralTileStreaming/Scripts/ProceduralScatter.cs
@@ -5,9 +5,11 @@ using Unity.Collections;
 using Unity.Entities;
 using Unity.Jobs;
 using Unity.Mathematics;
+using Unity.Rendering;
 using Unity.Transforms;
 using UnityEngine;
 using UnityEngine.Assertions;
+using UnityEngine.SocialPlatforms;
 
 public struct ProceduralScatterPrefab : IBufferElementData
 {
@@ -17,8 +19,7 @@ public struct ProceduralScatterPrefab : IBufferElementData
 
 public struct ProceduralScatterTile : ISharedComponentData
 {
-    public int3      Position;
-    public Entity    ScatterPrefab;
+    public Entity    TileEntity;
 }
 
 public struct ProceduralInstanceData
@@ -33,6 +34,7 @@ struct TempProceduralScatterPrefabTag : IComponentData
     
 }
 
+[ExecuteAlways]
 class ScatterStreamingSystem : JobComponentSystem
 {
     struct Stream
@@ -43,13 +45,15 @@ class ScatterStreamingSystem : JobComponentSystem
         
         public Entity SrcScatterPrefab;
         public Entity StreamingWorldScatterPrefab;
+        public Entity TileEntity;
     }
 
     Stream m_Stream;
     EntityQuery m_ScatterPrefabQuery;
     EntityQuery m_StreamingInstancesQuery;
+    EntityQuery m_MarkStaticFrozen;
     EntityQuery m_UnloadInstancesQuery;
-
+    
     [BurstCompile]
     struct SortBatchesJob : IJob
     {
@@ -72,7 +76,6 @@ class ScatterStreamingSystem : JobComponentSystem
         public ExclusiveEntityTransaction Transaction;
         public Entity                     ScatterPrefab;
         public NativeArray<ProceduralInstanceData> InstanceData;
-
         public void Execute()
         {
             var prefabs = new NativeArray<Entity>(Transaction.GetBuffer<ProceduralScatterPrefab>(ScatterPrefab).Reinterpret<Entity>().AsNativeArray(), Allocator.Temp);
@@ -95,8 +98,31 @@ class ScatterStreamingSystem : JobComponentSystem
                 // Place objects
                 for (int j = 0; j != batchCount; j++)
                 {
-                    Transaction.SetComponentData(instances[j], new Translation {Value = InstanceData[j + beginBatch].Position});
-                    Transaction.SetComponentData(instances[j], new Rotation {Value = InstanceData[j + beginBatch].Rotation});
+                    var instance = instances[j];
+
+                    var translation = InstanceData[j + beginBatch].Position;
+                    var rotation =  InstanceData[j + beginBatch].Rotation;
+                    
+                    if (Transaction.HasComponent(instance, ComponentType.ReadWrite<Translation>()))
+                        Transaction.SetComponentData(instance, new Translation { Value = Transaction.GetComponentData<Translation>(instance).Value + translation });
+                    if (Transaction.HasComponent(instance, ComponentType.ReadWrite<Rotation>()))
+                        Transaction.SetComponentData(instance, new Rotation { Value = math.mul(Transaction.GetComponentData<Rotation>(instance).Value, rotation) });
+                    
+                    if (Transaction.HasComponent(instance, ComponentType.ReadWrite<LinkedEntityGroup>()))
+                    {
+                        var linked = Transaction.GetBuffer<LinkedEntityGroup>(instance);
+                        foreach (var e in linked)
+                        {
+                            if (Transaction.HasComponent(e.Value, ComponentType.ReadWrite<LocalToWorld>()))
+                            {
+                                var localToWorld = Transaction.GetComponentData<LocalToWorld>(e.Value);
+                                var trs = float4x4.TRS(translation, rotation, new float3(1));
+                                localToWorld.Value = math.mul(trs, localToWorld.Value);
+                                
+                                Transaction.SetComponentData(e.Value, localToWorld);
+                            }
+                        }
+                    }
                 }
                 beginBatch = endBatch;
             }
@@ -127,10 +153,12 @@ class ScatterStreamingSystem : JobComponentSystem
         for (int i = 0; i != dstPrefabs.Length; i++)
         {
             var linkedEntities = EntityManager.GetBuffer<LinkedEntityGroup>(dstPrefabs[i]);
+
             for (int j = 0; j != linkedEntities.Length; j++)
             {
-                commands.AddComponent(linkedEntities[j].Value, default(TempProceduralScatterPrefabTag));
-                commands.AddComponent(linkedEntities[j].Value, default(Prefab));
+                var child = linkedEntities[j].Value;
+                commands.AddComponent(child, default(TempProceduralScatterPrefabTag));
+                commands.AddComponent(child, default(Prefab));
             }
         }
         commands.Playback(EntityManager);
@@ -146,6 +174,8 @@ class ScatterStreamingSystem : JobComponentSystem
         entityRemapping.Dispose();
         
         stream.StreamingWorld.EntityManager.RemoveComponent(stream.StreamingWorld.EntityManager.UniversalQuery, typeof(TempProceduralScatterPrefabTag));
+        stream.StreamingWorld.EntityManager.RemoveComponent<EditorRenderData>(stream.StreamingWorld.EntityManager.UniversalQuery);
+        stream.StreamingWorld.EntityManager.RemoveComponent<EntityGuid>(stream.StreamingWorld.EntityManager.UniversalQuery);
     }
 
     public bool ShouldGenerateTile
@@ -154,7 +184,7 @@ class ScatterStreamingSystem : JobComponentSystem
     }
     
     //@TODO: Warn about duplicate tiles
-    public JobHandle GenerateTileAsync(int3 tile, Entity scatterPrefab, NativeList<ProceduralInstanceData> instance, JobHandle dependency)
+    public JobHandle GenerateTileAsync(int3 tile, Entity tileEntity, Entity scatterPrefab, NativeList<ProceduralInstanceData> instance, JobHandle dependency)
     {
         if (!ShouldGenerateTile)
             throw new System.ArgumentException("GenerateTileAsync can only be called when ShouldGenerateTile returns true");
@@ -174,16 +204,22 @@ class ScatterStreamingSystem : JobComponentSystem
         
         m_Stream.StreamingWorld.EntityManager.ExclusiveEntityTransactionDependency = dependency;
         m_Stream.Tile = tile;
+        m_Stream.TileEntity = tileEntity;
         m_Stream.IsGeneratingTile = true;
 
         return dependency;
     }
     
-    public void UnloadTile(int3 tile, Entity scatterPrefab)
+    public bool UnloadTile(Entity tileEntity)
     {
-        m_UnloadInstancesQuery.SetSharedComponentFilter(new ProceduralScatterTile { Position = tile, ScatterPrefab = scatterPrefab});
+        if (m_Stream.IsGeneratingTile && m_Stream.TileEntity == tileEntity)
+            return false;
+        
+        m_UnloadInstancesQuery.SetSharedComponentFilter(new ProceduralScatterTile { TileEntity = tileEntity });
         EntityManager.DestroyEntity(m_UnloadInstancesQuery);
         m_UnloadInstancesQuery.ResetFilter();
+
+        return true;
     }
 
     protected override JobHandle OnUpdate(JobHandle inputDeps)
@@ -193,13 +229,19 @@ class ScatterStreamingSystem : JobComponentSystem
             m_Stream.StreamingWorld.EntityManager.EndExclusiveEntityTransaction();
 
             //@TODO: Move this to tile generation job
-            m_Stream.StreamingWorld.EntityManager.AddSharedComponentData(m_StreamingInstancesQuery, new ProceduralScatterTile { Position = m_Stream.Tile, ScatterPrefab = m_Stream.SrcScatterPrefab });
+            
+            //@TODO: Need some allocation scheme for FrozenRenderSceneTag to avoid conflicts...
+            var tileGUID = new Unity.Entities.Hash128(1U, (uint) m_Stream.Tile.x, (uint) m_Stream.Tile.y, (uint) m_Stream.Tile.z);
+            m_Stream.StreamingWorld.EntityManager.AddSharedComponentData(m_MarkStaticFrozen, new FrozenRenderSceneTag() { SceneGUID = tileGUID});
+            m_Stream.StreamingWorld.EntityManager.AddSharedComponentData(m_StreamingInstancesQuery, new ProceduralScatterTile { TileEntity = m_Stream.TileEntity });
             
             var entityRemapping = m_Stream.StreamingWorld.EntityManager.CreateEntityRemapArray(Allocator.TempJob);
             EntityManager.MoveEntitiesFrom(m_Stream.StreamingWorld.EntityManager, m_StreamingInstancesQuery, entityRemapping);
             entityRemapping.Dispose();
             
             m_Stream.IsGeneratingTile = false;
+
+            //Debug.Log($"Moved Tile: {tileGUID}");
         }
             
         return inputDeps;
@@ -208,16 +250,12 @@ class ScatterStreamingSystem : JobComponentSystem
     protected override void OnCreate()
     {       
         m_Stream.StreamingWorld = new World("StreamingWorld");
-
+        var streamingManager = m_Stream.StreamingWorld.EntityManager;
+        
         m_ScatterPrefabQuery = GetEntityQuery(new EntityQueryDesc 
         { 
             All = new[] { ComponentType.ReadOnly<TempProceduralScatterPrefabTag>() }, 
             Options = EntityQueryOptions.IncludePrefab | EntityQueryOptions.IncludeDisabled
-        });
-        m_StreamingInstancesQuery = m_Stream.StreamingWorld.EntityManager.CreateEntityQuery(new EntityQueryDesc 
-        { 
-            None = new[] { ComponentType.ReadOnly<ProceduralScatterPrefab>() },
-            Options = EntityQueryOptions.IncludeDisabled
         });
         
         m_UnloadInstancesQuery = GetEntityQuery(new EntityQueryDesc 
@@ -225,7 +263,18 @@ class ScatterStreamingSystem : JobComponentSystem
             All = new[] { ComponentType.ReadOnly<ProceduralScatterTile>() }, 
             Options = EntityQueryOptions.IncludePrefab | EntityQueryOptions.IncludeDisabled
         });
-        
+
+        m_MarkStaticFrozen = streamingManager.CreateEntityQuery(new EntityQueryDesc 
+        { 
+            All = new[] { ComponentType.ReadOnly<Static>() },
+            Options = EntityQueryOptions.IncludeDisabled
+        });
+        m_StreamingInstancesQuery = streamingManager.CreateEntityQuery(new EntityQueryDesc 
+        { 
+            None = new[] { ComponentType.ReadOnly<ProceduralScatterPrefab>() },
+            Options = EntityQueryOptions.IncludeDisabled
+        });
+
         RequireForUpdate(GetEntityQuery(ComponentType.ReadOnly<ProceduralScatterPrefab>()));
     }
